Time: 48 ms (8.47%), Space: 47.7 MB (99.25%) - LeetHub

+++ 1002-maximum-width-ramp/1002-maximum-width-ramp.cpp
@@ -25,7 +25,7 @@ public:
     int maxWidthRamp(vector<int>& nums) {
         int  maxWidth = 0;
         unsigned short i;
-        stack<int>stk;
+        stack<unsigned short>stk;
         stk.push(0);
         for ( i = 1 ; i < nums.size() ; i++ )
         {
